--- 03-next-travel/src/app/boards/detail/page.tsx
@@ -1,76 +1,0 @@
-"use client";
-
-import Button from '@/components/button';
-import styles from './styles.module.css';
-
-export default function BoardsDetail() {
-
-    return (
-        <div className={styles['main-content']}>
-            <div className={styles['page-container']}>
-                <h1 className={styles['page-title']}>살어리 살어리랏다 쳥산(靑山)애 살어리랏다멀위랑 ᄃᆞ래랑 먹고 쳥산(靑山)애 살어리랏다얄리얄리 얄랑셩 얄라리 얄라</h1>
-                <div className={styles['boards-detail-info']}>
-                    <div className={styles['boards-detail-profile-date']}>
-                        <div className={styles['boards-detail-profile']}>
-                            <img src="/assets/images/profileimg.png" className={styles['profile-image-small']} />홍길동
-                        </div>
-                        <div>2024.11.11</div>
-                    </div>
-                    <div className={styles['divider']}></div>
-                    <div className={styles['boards-deatil-button-group1']}>
-                        {/* 이 버튼들은 글로벌 CSS를 사용합니다. */}
-                        <Button className="simple-button" icon="/assets/icons/outline/link.svg" />
-                        <Button className="simple-button" icon="/assets/icons/outline/location.svg" />
-                    </div>
-                </div>
-                <div className={styles['detail-container']}>
-                    <img src="/assets/images/detailbeach.jpg" className={styles['detail-image1']} />
-                    살겠노라 살겠노라. 청산에 살겠노라.<br />
-                    머루랑 다래를 먹고 청산에 살겠노라.<br />
-                    얄리얄리 얄랑셩 얄라리 얄라<br />
-                    <br />
-                    우는구나 우는구나 새야. 자고 일어나 우는구나 새야.<br />
-                    너보다 시름 많은 나도 자고 일어나 우노라.<br />
-                    얄리얄리 얄라셩 얄라리 얄라<br />
-                    <br />
-                    갈던 밭(사래) 갈던 밭 보았느냐. 물 아래(근처) 갈던 밭 보았느냐<br />
-                    이끼 묻은 쟁기를 가지고 물 아래 갈던 밭 보았느냐.<br />
-                    얄리얄리 얄라셩 얄라리 얄라<br />
-                    <br />
-                    이럭저럭 하여 낮일랑 지내 왔건만<br />
-                    올 이도 갈 이도 없는 밤일랑 또 어찌 할 것인가.<br />
-                    얄리얄리 얄라셩 얄라리 얄라<br />
-                    <br />
-                    어디다 던지는 돌인가 누구를 맞히려던 돌인가.<br />
-                    미워할 이도 사랑할 이도 없이 맞아서 우노라.<br />
-                    얄리얄리 얄라셩 얄라리 얄라<br />
-                    <br />
-                    살겠노라 살겠노라. 바다에 살겠노라.<br />
-                    나문재, 굴, 조개를 먹고 바다에 살겠노라.<br />
-                    얄리얄리 얄라셩 얄라리 얄라<br />
-                    <br />
-                    가다가 가다가 듣노라. 에정지(미상) 가다가 듣노라.<br />
-                    사슴(탈 쓴 광대)이 솟대에 올라서 해금을 켜는 것을 듣노라.<br />
-                    얄리얄리 얄라셩 얄라리 얄라<br />
-                    <br />
-                    가다 보니 배불룩한 술독에 독한 술을 빚는구나.<br />
-                    조롱박꽃 모양 누룩이 매워 (나를) 붙잡으니 내 어찌 하리이까.[1]<br />
-                    얄리얄리 얄라셩 얄라리 얄라
-                    <img src="/assets/images/detailvideoimage.png" className={styles['detail-image2']} />
-                </div>
-                <div className={styles['boards-detail-action']}>
-                    <div className={styles['boards-deatil-button-group2']}>
-                        {/* 이 버튼들도 글로벌 CSS를 사용합니다. */}
-                        <Button className="simple-vertical-button" icon="/assets/icons/outline/bad.svg" text="24" />
-                        <Button className="simple-vertical-button red-button" icon="/assets/icons/outline/good.svg" text="12" />
-                    </div>
-                    <div className={styles['boards-deatil-button-group2']}>
-                        {/* 이 버튼들도 글로벌 CSS를 사용합니다. */}
-                        <Button className="white-button" icon="/assets/icons/outline/menu.svg" text="목록으로" />
-                        <Button className="white-button" icon="/assets/icons/outline/edit.svg" text="수정하기" />
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
-}
